Move RPA integration classes into RPA functional package
